added alert if trying to add coin already have added

+++ src/data/addNewCoin.js
@@ -2,14 +2,26 @@ import loadSavedCoins from './loadSavedCoins';
 
 const addNewCoin = (symbol, holdAmount) => {
     const newCoin = { symbol: symbol, holdAmount: holdAmount };
-    const currentCoins = localStorage.getItem('savedCoins');
+    const currentCoins = loadSavedCoins();
     if (currentCoins === null) {
         const arr = [newCoin];
         localStorage.setItem('savedCoins', JSON.stringify(arr));
     } else {
-        const currentCoins = loadSavedCoins();
-        const updatedArr = [...currentCoins, newCoin];
-        localStorage.setItem('savedCoins', JSON.stringify(updatedArr));
+        let alreadyOwnCoin = false;
+        for (var i = 0; i < currentCoins.length; i++) {
+            if (symbol === currentCoins[i].symbol) {
+                alreadyOwnCoin = true;
+            }
+        }
+
+        if (alreadyOwnCoin === false) {
+            const updatedArr = [...currentCoins, newCoin];
+            localStorage.setItem('savedCoins', JSON.stringify(updatedArr));
+        } else {
+            alert(
+                'You have already added this coin before! Edit hold amount or delete using edit button.'
+            );
+        }
     }
 };
